Refactor function that sync the schedule in another class
